requirements.txt
biopython==1.83
dill==0.3.8
gemmi==0.6.6
joblib==1.4.2
matplotlib==3.9.0
MDAnalysis==2.7.0
mmcif-pdbx==2.0.1
mmtf-python==1.1.3
multiprocess==0.70.16
numpy==1.26.4
pandas==2.2.2
pdb2pqr==3.6.2
pdbinf==0.0.4
prolif==2.0.3
propka==3.5.1
py3Dmol==2.1.0
rdkit==2023.9.6
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
seaborn==0.13.2